Add initial server and tampermonkey script
The server is just a websocket that broadcasts messages to all clients except
for the original sender.

The tampermonkey script works on HBO Max. It listens to the websocket and plays/pauses
the video on messages. When the space bar is pressed it sends its own play/pause
message on the websocket.